Split codes about handler to handler.go
