+++ main.go
@@ -21,35 +21,6 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 )
 
-type handler struct {
-}
-
-// Home is a handler method for /
-func (h *handler) Home() func(w http.ResponseWriter, r *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		resp, err := http.Get("https://lottery-dot-tenntenn-samples.appspot.com/available_lotteries")
-		if err != nil {
-			const status = http.StatusInternalServerError
-			http.Error(w, http.StatusText(status), status)
-			return
-		}
-		defer resp.Body.Close()
-
-		var lotteries []*Lottery
-		if err := json.NewDecoder(resp.Body).Decode(&lotteries); err != nil {
-			const status = http.StatusInternalServerError
-			http.Error(w, http.StatusText(status), status)
-			return
-		}
-
-		if err := listTmpl.Execute(w, lotteries); err != nil {
-			const status = http.StatusInternalServerError
-			http.Error(w, http.StatusText(status), status)
-			return
-		}
-	}
-}
-
 func main() {
 
 	db, err := sql.Open("sqlite3", "database.db")
